requirements.txt
matplotlib==3.7.2
numpy==1.25.2
pandas==1.5.3
Pillow==11.3.0
scikit_learn==1.3.1
seaborn==0.11.2
torch==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.66.1
pytorch-fid==0.3.0

